+++ src/tracer/bxdfs/dielectric.cpp
@@ -7,8 +7,8 @@ using namespace vec_helpers;
 
 namespace {
 // helpers
-static double FrDielectric(double cos0, double eta) {
-  cos0 = std::clamp<double>(cos0, -1, 1);
+static Float FrDielectric(Float cos0, Float eta) {
+  cos0 = std::clamp<Float>(cos0, -1, 1);
   // Potentially flip interface orientation for Fresnel equations
   if (cos0 < 0) {
     eta = 1 / eta;
@@ -16,20 +16,20 @@ static double FrDielectric(double cos0, double eta) {
   }
 
   // Compute $\cos\,\theta_\roman{t}$ for Fresnel equations using Snell's law
-  double sin2Theta_o = 1 - Sqr(cos0);
-  double sin2Theta_t = sin2Theta_o / Sqr(eta);
+  Float sin2Theta_o = 1 - Sqr(cos0);
+  Float sin2Theta_t = sin2Theta_o / Sqr(eta);
   if (sin2Theta_t >= 1)
     return 1.f;
-  double cosTheta_t = std::sqrt(1 - sin2Theta_t);
+  Float cosTheta_t = std::sqrt(1 - sin2Theta_t);
 
-  double r_parl = (eta * cos0 - cosTheta_t) / (eta * cos0 + cosTheta_t);
-  double r_perp = (cos0 - eta * cosTheta_t) / (cos0 + eta * cosTheta_t);
+  Float r_parl = (eta * cos0 - cosTheta_t) / (eta * cos0 + cosTheta_t);
+  Float r_perp = (cos0 - eta * cosTheta_t) / (cos0 + eta * cosTheta_t);
   return (Sqr(r_parl) + Sqr(r_perp)) / 2;
 }
 
 }  // namespace
 
-Dielectric::Dielectric(double index, TrowbridgeReitzDistribution mfdist)
+Dielectric::Dielectric(Float index, TrowbridgeReitzDistribution mfdist)
     : BxDF(BxDFBits::Transmission), eta(index), mfdist_(std::move(mfdist)) {
   if (eta != 1)
     SetFlags(GetFlags() | BxDFBits::Reflection);
@@ -41,10 +41,10 @@ Color Dielectric::f(const Vector3& wi, const Vector3& wo) const {
   if (mfdist_.EffectivelySmooth() || eta == 1)
     return Color(0);
 
-  double cos0_i = CosTheta(-wi), cos0_o = CosTheta(wo);
+  Float cos0_i = CosTheta(-wi), cos0_o = CosTheta(wo);
   bool is_reflect = cos0_i * cos0_o > 0;
 
-  double etap = 1;
+  Float etap = 1;
   if (!is_reflect)
     etap = cos0_i > 0 ? eta : (1.0 / eta);
 
@@ -57,14 +57,14 @@ Color Dielectric::f(const Vector3& wi, const Vector3& wo) const {
   if (wm.Dot(wo) * cos0_o < 0 || wm.Dot(-wi) * cos0_i < 0)
     return Color(0);
 
-  double F = FrDielectric(wm.Dot(-wi), eta);
+  Float F = FrDielectric(wm.Dot(-wi), eta);
   if (is_reflect) {
-    double c =
+    Float c =
         mfdist_.D(wm) * mfdist_.G(wi, wo) * F / std::abs(4 * cos0_i * cos0_o);
     return Color(c);
   } else {
-    double denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap) * cos0_i * cos0_o;
-    double ft = mfdist_.D(wm) * (1 - F) * mfdist_.G(wi, wo) *
+    Float denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap) * cos0_i * cos0_o;
+    Float ft = mfdist_.D(wm) * (1 - F) * mfdist_.G(wi, wo) *
                 std::abs(wm.Dot(wo) * wm.Dot(-wi) / denom);
 
     ft /= Sqr(etap);
@@ -72,14 +72,14 @@ Color Dielectric::f(const Vector3& wi, const Vector3& wo) const {
   }
 }
 
-double Dielectric::pdf(const Vector3& wi, const Vector3& wo) const {
+Float Dielectric::pdf(const Vector3& wi, const Vector3& wo) const {
   if (mfdist_.EffectivelySmooth() || eta == 1)
     return 0;
 
-  double cos0_i = CosTheta(-wi), cos0_o = CosTheta(wo);
+  Float cos0_i = CosTheta(-wi), cos0_o = CosTheta(wo);
   bool is_reflect = cos0_i * cos0_o > 0;
 
-  double etap = 1;
+  Float etap = 1;
   if (!is_reflect)
     etap = cos0_i > 0 ? eta : (1.0 / eta);
 
@@ -92,15 +92,15 @@ double Dielectric::pdf(const Vector3& wi, const Vector3& wo) const {
   if (wm.Dot(wo) * cos0_o < 0 || wm.Dot(-wi) * cos0_i < 0)
     return 0;
 
-  double pr = FrDielectric(wm.Dot(-wi), eta);
-  double pt = 1.0 - pr;
+  Float pr = FrDielectric(wm.Dot(-wi), eta);
+  Float pt = 1.0 - pr;
 
-  double pdf = 0;
+  Float pdf = 0;
   if (is_reflect) {
     pdf = pr * mfdist_.PDF(-wi, wm) / (4 * std::abs(wm.Dot(wi)));
   } else {
-    double denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap);
-    double dwm_dwo = std::abs(wm.Dot(wo)) / denom;
+    Float denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap);
+    Float dwm_dwo = std::abs(wm.Dot(wo)) / denom;
     pdf = mfdist_.PDF(-wi, wm) * dwm_dwo * pt;
   }
   return pdf;
@@ -108,13 +108,13 @@ double Dielectric::pdf(const Vector3& wi, const Vector3& wo) const {
 
 std::optional<bxdfSample> Dielectric::Sample_f(const Vector3& wi) const {
   bool entering = wi.y() < 0;
-  double etaI = entering ? 1.0 : eta;
-  double etaT = entering ? eta : 1.0;
-  double etaRel = etaI / etaT;
+  Float etaI = entering ? 1.0 : eta;
+  Float etaT = entering ? eta : 1.0;
+  Float etaRel = etaI / etaT;
 
   if (eta == 1 || mfdist_.EffectivelySmooth()) {
-    double pr = FrDielectric(CosTheta(-wi), eta);
-    double pt = 1 - pr;
+    Float pr = FrDielectric(CosTheta(-wi), eta);
+    Float pt = 1 - pr;
 
     if (random_uniform_01() < pr) {
       Vector3 wo(wi.x(), -wi.y(), wi.z());
@@ -133,31 +133,31 @@ std::optional<bxdfSample> Dielectric::Sample_f(const Vector3& wi) const {
     }
   } else {
     Vector3 wm = mfdist_.Sample_wm(-wi);
-    double pr = FrDielectric(wm.Dot(-wi), eta);
-    double pt = 1.0 - pr;
+    Float pr = FrDielectric(wm.Dot(-wi), eta);
+    Float pt = 1.0 - pr;
 
     if (random_uniform_01() < pr) {
       Vector3 wo = Reflect(wi, wm).Normalized();
       if (!SameHemisphere(wi, wo))
         return std::nullopt;
 
-      double pdf = mfdist_.PDF(-wi, wm) / (4 * std::abs(wm.Dot(wi))) * pr;
-      double f = mfdist_.D(wm) * mfdist_.G(wi, wo) * pr /
+      Float pdf = mfdist_.PDF(-wi, wm) / (4 * std::abs(wm.Dot(wi))) * pr;
+      Float f = mfdist_.D(wm) * mfdist_.G(wi, wo) * pr /
                  (4 * CosTheta(wo) * CosTheta(-wi));
       return bxdfSample(Color(f), wo, pdf,
                         BxDFBits::Glossy | BxDFBits::Reflection);
     } else {
       Vector3 wo;
-      double etap = etaRel;
+      Float etap = etaRel;
       bool ok = Refract(wi, wm, etap, &wo);
       if (!ok || SameHemisphere(wi, wo) || wo.y() == 0)
         return std::nullopt;
 
       wo = wo.Normalized();
-      double denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap);
-      double dwm_dwo = std::abs(wm.Dot(wo)) / denom;
-      double pdf = mfdist_.PDF(-wi, wm) * dwm_dwo * pt;
-      double ft = pt * mfdist_.D(wm) * mfdist_.G(wi, wo) / denom *
+      Float denom = Sqr(wm.Dot(wo) + wm.Dot(-wi) / etap);
+      Float dwm_dwo = std::abs(wm.Dot(wo)) / denom;
+      Float pdf = mfdist_.PDF(-wi, wm) * dwm_dwo * pt;
+      Float ft = pt * mfdist_.D(wm) * mfdist_.G(wi, wo) / denom *
                   std::abs(wm.Dot(wo) * wm.Dot(-wi) /
                            (AbsCosTheta(wo) * AbsCosTheta(wi)));
