+++ src/tracer/hit_record.hpp
@@ -10,12 +10,12 @@ struct HitRecord {
   // note: everythin in local coordinate system
   bool hits = false;
   Primitive const* primitive = nullptr;
-  double time;
+  Float time;
   Point3 position;
   Point2 uv;
   Normal normal;  // outward normal
 
-  static constexpr HitRecord Create(double time,
+  static constexpr HitRecord Create(Float time,
                                     Point3 pos,
                                     Point2 uv,
                                     Normal n) {
