+++ src/tracer/shapes/2d/triangle.hpp
@@ -7,11 +7,11 @@ class Triangle : public IShape {
   Triangle(Point3 o, Point3 a, Point3 b);
 
   AABB GetBbox() const override;
-  HitRecord Hit(const Ray&, const Interval<double>&) const override;
+  HitRecord Hit(const Ray&, const Interval<Float>&) const override;
   ShapeSample Sample() const override;
-  double Area() const override;
+  Float Area() const override;
 
  private:
   MatrixTransformation trans_;
-  double area_;
+  Float area_;
 };
